move unused components
