+++ src/App.tsx
@@ -12,7 +12,6 @@ import ResetPassword from './pages/ResetPassword';
 import { NotFoundPage } from './pages/404';
 import { AppState } from './store';
 import style from './style';
-import { MitPreprocessDemo } from './pages/mit/MitPreprocessDemo';
 import { routes } from './pages/routes';
 
 // 公共的页面
@@ -21,7 +20,6 @@ const publicPaths = [
   routes.login,
   routes.signUp,
   routes.resetPassword,
-  routes.mit.preprocessDemo,
 ] as readonly string[];
 
 const App: React.FC = () => {
@@ -168,9 +166,6 @@ const App: React.FC = () => {
           </Route>
           <Route path={routes.dashboard.$}>
             <Dashboard />
-          </Route>
-          <Route path={routes.mit.preprocessDemo}>
-            <MitPreprocessDemo />
           </Route>
           {userIsAdmin && (
             <Route path={routes.admin}>
