+++ src/services/moeflow_companion/TranslateCompanion.tsx
@@ -2,8 +2,8 @@ import { FC } from '@/interfaces';
 import { RefObject, useRef, useState } from 'react';
 import { FilePond } from 'react-filepond';
 import { css } from '@emotion/core';
-import { Button } from '../Button';
-import { createMoeflowProjectZip, LPFile } from './moeflow-packager';
+import { Button } from '@/components/Button';
+import { createMoeflowProjectZip, LPFile } from '../labelplus_packager';
 import { FailureResults } from '@/apis';
 import { measureImgSize } from '@jokester/ts-commonutil/lib/web/measure-img';
 import { clamp } from 'lodash-es';
